Note: Apologies for the late submission sir... I understand the penalties for late submission.
I was stressed before the deadline but I found out that (if) conditions are needed to filter out the different values from MainActivity.java to SecondActivity.java (supposed to be AnswersActivity but lazy change name)
The if conditions are whether variable value equals to intent "value:" from the unique envelopes of content. And the text will display based on the content.

activity_main.xml change txtPassDub's layout_width to wrap_content so that gravity works.

I still don't know what the new vcs.xml is, but it seems like it will always be added for new commit